consensus: make `Header` gas limit u64 (#1333)
* consensus: make Header gas limit u64

* fix clippy

* make gas used u64

* Update Cargo.toml

---------

+++ crates/network-primitives/src/traits.rs
@@ -153,7 +153,7 @@ pub trait HeaderResponse {
     fn coinbase(&self) -> Address;
 
     /// Gas limit of the block
-    fn gas_limit(&self) -> u128;
+    fn gas_limit(&self) -> u64;
 
     /// Mix hash of the block
     ///
@@ -392,7 +392,7 @@ impl<T: HeaderResponse> HeaderResponse for WithOtherFields<T> {
         self.inner.coinbase()
     }
 
-    fn gas_limit(&self) -> u128 {
+    fn gas_limit(&self) -> u64 {
         self.inner.gas_limit()
     }
